Combine edit and create
